+++ app/Port/Middleware/Http/Localization.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Port\Middleware\Middlewares\Http;
+namespace App\Port\Middleware\Http;
 
 use Closure;
 use Illuminate\Foundation\Application;
